make one component and make one test file for that component. remove the two components and associated files the new component replaces
