+++ ui/app/components/secret-engine/configure-wif.hbs
@@ -3,11 +3,28 @@
   SPDX-License-Identifier: BUSL-1.1
 ~}}
 
-<form {{on "submit" (perform this.submitForm)}} aria-label="configure azure credentials" data-test-configure-form>
+<form
+  {{on "submit" (perform this.submitForm)}}
+  aria-label={{concat "configure" @type " credentials"}}
+  data-test-configure-form
+>
   <NamespaceReminder @mode="save" @noun="configuration" />
   <MessageError @errorMessage={{this.errorMessage}} />
+  {{! AWS specific note and section header }}
+  {{#if (eq @type "aws")}}
+    <div class="box is-fullwidth is-shadowless is-marginless">
+      <p class="has-text-grey-dark">
+        Note: the client uses the official AWS SDK and will use the specified credentials, environment credentials, shared
+        file credentials, or IAM role/ECS task credentials in that order.
+      </p>
+    </div>
+    <h2 class="title is-5 has-bottom-padding-s has-top-margin-l" data-test-access-title>
+      Access to AWS
+    </h2>
+  {{/if}}
+
   <div class="box is-fullwidth is-sideless">
-    {{! accessType can be "azure" or "wif" - since WIF is an enterprise only feature we default to "azure" for community users and only display those related form fields. }}
+    {{! Only enterprise users can select a specific access type: "account" or "wif" }}
     {{#if this.version.isEnterprise}}
       <fieldset class="field form-fieldset" id="protection" data-test-access-type-section>
         <legend class="is-label">Access Type</legend>
@@ -15,22 +32,24 @@
           {{#if this.disableAccessType}}
             You cannot edit Access Type if you have already saved access credentials.
           {{else}}
-            Choose the way to configure access to Azure. Access can be configured either using an Azure account or with the
-            Plugin Workload Identity Federation (WIF).
+            Choose the way to configure access to
+            {{@displayName}}. Access can be configured either using an
+            {{@displayName}}
+            account credentials or with the Plugin Workload Identity Federation (WIF).
           {{/if}}
         </p>
         <div>
           <RadioButton
-            id="access-type-azure"
-            name="azure"
+            id={{concat "access-type-" @type}}
+            name="account"
             class="radio"
-            data-test-access-type="azure"
-            @value="azure"
+            data-test-access-type={{@type}}
+            @value="account"
             @groupValue={{this.accessType}}
-            @onChange={{fn this.onChangeAccessType "azure"}}
+            @onChange={{fn this.onChangeAccessType "account"}}
             @disabled={{this.disableAccessType}}
           />
-          <label for="access-type-azure">Azure account credentials</label>
+          <label for={{concat "access-type-" @type}}>{{@displayName}} account credentials</label>
           <RadioButton
             id="access-type-wif"
             name="wif"
@@ -46,21 +65,34 @@
       </fieldset>
     {{/if}}
     {{#if (eq this.accessType "wif")}}
-      {{! WIF Fields }}
+      {{! if access type is "wif" display Issuer and WIF fields }}
       {{#each @issuerConfig.displayAttrs as |attr|}}
         <FormField @attr={{attr}} @model={{@issuerConfig}} />
       {{/each}}
       <FormFieldGroups @model={{@model}} @mode={{if @model.isNew "create" "edit"}} @groupName="fieldGroupsWif" />
     {{else}}
-      {{! Azure Account Fields }}
+      {{! otherwise display account credential fields }}
       <FormFieldGroups
         @model={{@model}}
         @mode={{if @model.isNew "create" "edit"}}
         @useEnableInput={{true}}
-        @groupName="fieldGroupsAzure"
+        @groupName="fieldGroupsAccount"
       />
     {{/if}}
+    {{! secondModel fields should show regardless of the vault version or what access type is selected }}
+    {{#if @secondModel}}
+      <h2 class="title is-5 has-bottom-padding-s has-top-margin-l" data-test-second-model-title>
+        {{or @secondModel.modelNameDisplay "Additional"}}
+        Configuration
+      </h2>
+      <div class="box is-fullwidth is-sideless is-bottomless">
+        {{#each @secondModel.formFields as |attr|}}
+          <FormField @attr={{attr}} @model={{@secondModel}} @modelValidations={{this.modelValidations}} />
+        {{/each}}
+      </div>
+    {{/if}}
   </div>
+
   <Hds::ButtonSet>
     <Hds::Button
       @text="Save"
