+++ ui/app/templates/vault/cluster/secrets/backend/configuration/edit.hbs
@@ -15,7 +15,7 @@
   <p.levelLeft>
     <h1 class="title is-3" data-test-backend-configure-title={{this.model.type}}>
       Configure
-      {{get (options-for-backend this.model.type) "displayName"}}
+      {{this.model.displayName}}
     </h1>
   </p.levelLeft>
 </PageHeader>
@@ -28,19 +28,16 @@
   </ToolbarActions>
 </Toolbar>
 
-{{#if (eq this.model.type "aws")}}
-  <SecretEngine::ConfigureAws
-    @leaseConfig={{this.model.aws-lease-config}}
-    @rootConfig={{this.model.aws-root-config}}
-    @issuerConfig={{this.model.identity-oidc-config}}
+{{#if (eq this.model.type "ssh")}}
+  <SecretEngine::ConfigureSsh @model={{this.model.first-model}} @id={{this.model.id}} />
+  {{! This check is preventive. As of writing all other engines using this route——but ssh——are wif engines }}
+{{else if this.model.isWifEngine}}
+  <SecretEngine::ConfigureWif
     @backendPath={{this.model.id}}
-  />
-{{else if (eq this.model.type "azure")}}
-  <SecretEngine::ConfigureAzure
-    @model={{this.model.azure-config}}
-    @backendPath={{this.model.id}}
+    @displayName={{this.model.displayName}}
+    @type={{this.model.type}}
+    @model={{this.model.first-model}}
+    @secondModel={{this.model.second-model}}
     @issuerConfig={{this.model.identity-oidc-config}}
   />
-{{else if (eq this.model.type "ssh")}}
-  <SecretEngine::ConfigureSsh @model={{this.model.ssh-ca-config}} @id={{this.model.id}} />
 {{/if}}
