+++ ui/types/vault/models/secret-engine/config.d.ts
@@ -6,22 +6,43 @@
 import type Model from '@ember-data/model';
 import type { ModelValidations } from 'vault/app-types';
 
-export default class AzureConfig extends Model {
+export default class SecretEngineConfig extends Model {
   backend: string;
+  type: string;
+  // aws
+  leaseMax: any;
+  lease: any;
+  accessKey: any;
+  secretKey: any;
+  roleArn: any;
+  region: any;
+  iamEndpoint: any;
+  stsEndpoint: any;
+  maxRetries: any;
+  // azure
   subscriptionId: string | undefined;
   tenantId: string | undefined;
   clientId: string | undefined;
   clientSecret: string | undefined;
-  identityTokenAudience: string | undefined;
-  identityTokenTtl: any;
   environment: string | undefined;
   rootPasswordTtl: string | undefined;
+  // gcp
+  credentials: string | undefined;
+  ttl: any;
+  maxTtl: any;
+  secretAccountEmail: string | undefined;
+  modelNameDisplay: string | undefined;
+  // wif
+  identityTokenAudience: string | undefined;
+  identityTokenTtl: any;
 
   get displayAttrs(): any;
   get isWifPluginConfigured(): boolean;
-  get isAzureAccountConfigured(): boolean;
+  get isAccountPluginConfigured(): boolean;
   get fieldGroupsWif(): any;
   get fieldGroupsAzure(): any;
+  get fieldGroupsGcp(): any;
+  get fieldGroupsIam(): any;
   formFieldGroups(accessType?: string): {
     [key: string]: string[];
   }[];
@@ -31,4 +52,7 @@ export default class AzureConfig extends Model {
   isNew: boolean;
   save(): void;
   unloadRecord(): void;
+  destroyRecord(): void;
+  rollbackAttributes(): void;
+  validate(): ModelValidations;
 }
